fix: removed custom metrics from signal preview, replaced by 'Metrics'
